Ignore failing tests, see #81
